refactor: alterar tipo dos IDs de int para Long nas entidades e repositórios

+++ src/main/java/com/controlador/presenca/controllers/UsuarioController.java
@@ -22,7 +22,7 @@ public class UsuarioController {
     }
 
     @GetMapping(path = "/{id}")
-    public Optional<Usuario> getUsuarioById(@PathVariable int id) {
+    public Optional<Usuario> getUsuarioById(@PathVariable Long id) {
         return usuarioRepository.findById(id);
     }
 }
